Fix conversion warnings in libdispatch

+++ libdispatch/ncxcache.c
@@ -230,7 +230,7 @@ ncxcachenew(size_t leaflen, NCxcache** cachep)
     cache = calloc(1,sizeof(NCxcache));
     if(cache == NULL)
         {stat = NC_ENOMEM; goto done;}
-    cache->map = ncexhashnew(leaflen);
+    cache->map = ncexhashnew((int)leaflen);
     if(cache->map == NULL)
         {stat = NC_ENOMEM; goto done;}
     cache->lru.next = &cache->lru;
